Moved date utils into Utils
