+++ src/components/StackedDate.tsx
@@ -1,5 +1,7 @@
 import * as React from "react";
 
+import DateUtil from "../util/DateUtil";
+
 const stackedDateStyles = require("./StackedDate.css");
 
 interface IStackedDateProps {
@@ -8,27 +10,6 @@ interface IStackedDateProps {
 
 export default class StackedDate extends React.Component<IStackedDateProps, {}> {
 
-  public getDay(date: Date): string {
-    const options = {
-      day: "2-digit",
-    };
-    return date.toLocaleDateString(undefined, options);
-  }
-
-  public getMonth(date: Date): string {
-    const options = {
-      month: "long",
-    };
-    return date.toLocaleDateString(undefined, options);
-  }
-
-  public getYear(date: Date): string {
-    const options = {
-      year: "numeric",
-    };
-    return date.toLocaleDateString(undefined, options);
-  }
-
   public render() {
     return (
       <div className={stackedDateStyles.stackedDate}>
@@ -36,17 +17,17 @@ export default class StackedDate extends React.Component<IStackedDateProps, {}> 
         <div className={stackedDateStyles.stackedDateTop}>
 
           <div className={stackedDateStyles.stackedDateDay}>
-            {this.getDay(this.props.timestamp)}
+            {DateUtil.getDay(this.props.timestamp)}
           </div>
 
           <div className={stackedDateStyles.stackedDateYear}>
-            {this.getYear(this.props.timestamp)}
+            {DateUtil.getYear(this.props.timestamp)}
           </div>
 
         </div>
 
         <div className={stackedDateStyles.stackedDateMonth}>
-          {this.getMonth(this.props.timestamp)}
+          {DateUtil.getMonth(this.props.timestamp)}
         </div>
       </div>
     );
